Deal with inspection/cleaning flow issue
Before this it was possible for a node that failed cleaning to succeed
introspection and never be cleaned. This removes that possibility
by returning the node to a clean_wait state right away at which point
the inspection workflow will skip it and it will be picked up on the next
round.