Fix another clean/inspection edge case
The more we dig into the script the more it becomes clear that it
needs a proper redesign instead of trying to hack in cleaning retry
in the way it's done now.

Update the public subnet with a /22 cidr and allocation pool.

Adding four new variables for the subnet configuration.

Making the subnet range default to /16 cidr changing pool end.

My testing shows I can not use 255, making it 254 instead.

Matching the netmask to the cidr as was requested in a review comment.