Split introspection from undercloud install role
For OSP13 we should rely on bulk introspection due to the numerous
improvements made.  In order to reduce the complexity of the
undercloud install role I propose we split introspection out of the
undercloud install role.

This also separates bulk introspection from the introspection with retry
script in order to simplify the flow.  This also allows you to run either
bulk or the script manually as the script is always laid down rather than
templated.

This does complicate vars related to introspection so I am open to any
ideas to reduce that complexity.