Changes for OSP13
ostag requires several fixes to work correctly with OSP13, rather
than manage the dependencies across several repos, it is easier to
create a python script that interacts directly with the Ironic API
in order to pin nodes as we need.  There are also some Ansible
logic issues fixed such as variables not evaluating correctly
without a filter (Ex. |bool)