Raise Nova Max Scheduler Attempts
In the staging environment, we have a limited set of compute nodes
in which it is possible to encounter a race condition where as an
instance requiring PCI passthrough NVMe hits MaxRetriesExceeded.

Raising the number of possible attempts in order to increase likely
hood of successful instance booting in the staging environment with
PCI passthrough.